perfmemory-mcp/pyproject.toml
[project]
name = "perfmemory-mcp"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastmcp>=2.13.1",
  "psycopg2-binary>=2.9.11",
  "pgvector>=0.4.2",
  "openai>=2.30.0",
  "httpx>=0.28.1",
  "python-dotenv>=1.2.1",
  "numpy>=2.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services", "utils"]
include = ["perfmemory.py"]

[tool.hatch.build.targets.sdist]
include = [
  "services",
  "utils",
  "perfmemory.py",
  "pyproject.toml",
  "README.md",
  ".env.example",
]
